requirements.txt
pytorch-lightning
torchmetrics
wandb
loguru
segmentation-models-pytorch